requirements.txt
requests>=2.31.0
rich>=13.7.0
pathvalidate>=3.2.0
python-dotenv>=1.0.0
